Fix sample rate difference in tests; remove flaky server verification in tests
