+++ FunnelAITests/Integration/DeepgramClientTests.swift
@@ -13,9 +13,6 @@ class DeepgramClientTests: XCTestCase {
         // Initialize client
         client = DeepgramClient(serverBaseURL: "http://localhost:8000")
 
-        // Verify server is running
-        try await verifyServerIsRunning()
-
         // Get test audio file from the test bundle
         guard let url = Bundle(for: type(of: self)).url(
             forResource: "sample-recording-mary-had-lamb",
@@ -24,62 +21,41 @@ class DeepgramClientTests: XCTestCase {
             XCTFail("Could not find test audio file 'sample-recording-mary-had-lamb.m4a' in test bundle")
             return
         }
-        
+
         testAudioURL = url
         print("📦 Using audio file from test bundle: \(url)")
     }
-
-    private func verifyServerIsRunning() async throws {
-        let serverURL = URL(string: "http://localhost:8000")!
-        let request = URLRequest(url: serverURL)
-        
-        do {
-            let (_, response) = try await URLSession.shared.data(for: request)
-            guard let httpResponse = response as? HTTPURLResponse,
-                  httpResponse.statusCode == 200
-            else {
-                print("❌ Server returned status: \((response as? HTTPURLResponse)?.statusCode ?? -1)")
-                XCTFail("Server is not running at \(serverURL). Please start the server with 'cd server && deno task dev'")
-                throw XCTSkip("Server not running")
-            }
-            print("✅ Server is running at \(serverURL)")
-        } catch {
-            print("❌ Cannot connect to server: \(error)")
-            XCTFail("Cannot connect to server at \(serverURL): \(error)")
-            throw XCTSkip("Server not running")
-        }
-    }
-
 
     /// Test streaming audio data using DeepgramClient
     func testStreamAudioDataWithDeepgramClient() async throws {
         print("\n🎤 === DEEPGRAM CLIENT AUDIO STREAMING TEST START ===")
 
-        // Load audio data
+        // Load audio data (convertToPCM converts to 16kHz)
         let audioFile = try AVAudioFile(forReading: testAudioURL)
         let pcmData = try audioFile.convertToPCM()
-        print("🎵 Loaded \(pcmData.count) bytes of PCM data")
+        print("🎵 Loaded \(pcmData.count) bytes of PCM data at 16kHz")
 
         // Create a custom provider that streams in smaller chunks
         let chunkSize = 8000 // 0.5 seconds at 16kHz
         var offset = 0
         var chunkCount = 0
 
-        let processedRecording = try await client.streamRecording {
+        // Stream with explicit sample rate of 16000 Hz to match our converted audio
+        let processedRecording = try await client.streamRecording(sampleRate: 16000) {
             guard offset < pcmData.count else {
                 print("✅ Finished streaming \(chunkCount) chunks")
                 return nil
             }
-            
+
             let chunkEnd = min(offset + chunkSize, pcmData.count)
             let chunk = pcmData[offset ..< chunkEnd]
             offset = chunkEnd
             chunkCount += 1
-            
+
             if chunkCount % 5 == 0 {
                 print("📤 Streamed chunk \(chunkCount) (\(offset)/\(pcmData.count) bytes)")
             }
-            
+
             return chunk
         }
 
@@ -110,7 +86,8 @@ class DeepgramClientTests: XCTestCase {
             transcriptLower.contains("mary") || transcriptLower.contains("lamb"),
             "Transcript should contain words from Mary Had a Little Lamb"
         )
-        
+
         print("✅ DeepgramClient test completed successfully!")
     }
 }
+
